requirements.txt
beautifulsoup4==4.12.3
blinker==1.7.0
click==8.1.7
flask==3.0.2
google==3.0.0
importlib-metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mpegdash==0.4.0
protobuf==4.25.3
PyJWT==2.8.0
soupsieve==2.5
werkzeug==3.0.1
zipp==3.17.0
